pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mobile_clusters"
version = "0.1.2"
description = "A package to create MCL clusters"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Muhammad Sulman", email = "sulmanmu40@gmail.com"}
]

dependencies = [
    "pandas>=2.0.0",
    "biopython>=1.84",
    "tqdm>=2.0.0",
    "matplotlib>=3.0.0"]

[project.scripts]
mobile_clusters = "scripts.main:main"

[tool.setuptools]
packages = ["scripts"]
